registerFunctions seperated from register/route.ts file
